+++ models/settings/MotionType.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace app\models\settings;
+
+class MotionType
+{
+    use JsonConfigTrait;
+
+    /** @var string */
+    public $pdfIntroduction = '';
+    public $cssIcon = '';
+
+    /** @var bool */
+    public $layoutTwoCols = false;
+}