+++ models/settings/Site.php
@@ -11,15 +11,9 @@ class Site
 
     /** @var bool */
     public $showAntragsgruenAd  = true;
-    public $forceLogin          = false; // @TODO Delete this setting after migration is done
-    public $managedUserAccounts = false; // @TODO Delete this setting after migration is done
 
     /** @var int[] */
     public $loginMethods = [0, 1, 3];
-
-    /** @var null|string */
-    public $emailReplyTo  = null; // @TODO Delete this setting after migration is done
-    public $emailFromName = null; // @TODO Delete this setting after migration is done
 
     const LOGIN_STD        = 0;
     const LOGIN_WURZELWERK = 1;