+++ views/amendment/LayoutHelper.php
@@ -35,7 +35,7 @@ class LayoutHelper
             \Yii::t('amend', 'amendment_for_prefix')
         );
 
-        $intro                    = explode("\n", $amendment->getMyConsultation()->getSettings()->pdfIntroduction);
+        $intro                    = explode("\n", $amendment->getMyMotionType()->getSettings()->pdfIntroduction);
         $content->introductionBig = $intro[0];
         if (count($intro) > 1) {
             array_shift($intro);