+++ views/motion/LayoutHelper.php
@@ -90,8 +90,8 @@ class LayoutHelper
     {
         $hasAgenda                = ($motion->agendaItem !== null);
         $content                  = new Content();
-        $content->template        = $motion->motionType->texTemplate->texContent;
-        $intro                    = explode("\n", $motion->getMyConsultation()->getSettings()->pdfIntroduction);
+        $content->template        = $motion->getMyMotionType()->texTemplate->texContent;
+        $intro                    = explode("\n", $motion->getMyMotionType()->getSettings()->pdfIntroduction);
         $content->introductionBig = $intro[0];
         $content->titlePrefix     = $motion->titlePrefix;
         $content->titleLong       = $motion->getTitleWithPrefix();
@@ -117,7 +117,7 @@ class LayoutHelper
         }
 
         foreach ($motion->getSortedSections(true) as $section) {
-            $isRight = ($section->isLayoutRight() && $motion->motionType->layoutTwoCols);
+            $isRight = ($section->isLayoutRight() && $motion->motionType->getSettings()->layoutTwoCols);
             $section->getSectionType()->printMotionTeX($isRight, $content, $motion->getMyConsultation());
         }